requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-aiplatform
google-cloud-bigquery